attestation_verifier/server/requirements.txt
pycryptodome==3.9.9
requests==2.28.1
